requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
langchain==0.2.1
langchain-openai==0.1.8
langchain-community==0.2.1
openai==1.30.1
faiss-cpu==1.12.0
sentence-transformers==2.7.0
python-docx==1.1.0
python-multipart==0.0.9
aiofiles==23.2.1
tiktoken==0.7.0
numpy==2.0.0
